PXC-3417: Fixed assertion wsrep checkpoint assertion when DML executed in TOI
Problem:
We have the case when REPLACE INTO ... SELECT or INSERT INTO ... SELECT
is executed as TOI. This is done by	pt-table-checksum tool.
In such case statement execution causes InnoDB commit and storing of
xid_seqno (wsrep_apply_cb) and then because of TOI, storing of xid_seqno
is requested again from wsrep_commit_cb with the same xid_seqno.

Solution:
Allow current and new values be the same, without introducing new flags
and logic to prevent double storing	of the same value. This solution is
backport of current assertion condition in 8.0 branch.