PXC-4341: Executing prepared statement can abort node after FLUSH TABLES
https://perconadev.atlassian.net/browse/PXC-4341

Problem
-------
Transient errors during the execution of the DDL in a prepared statement
mode can trigger inconsistency voting resulting in node eviction from
the cluster.

Analysis & Fix
--------------
It can sometimes happen that the DDL executed as a prepared statement
may fail with transient errors like ER_NEED_REPREPARE if the table
metadata was changed between PREPARE and EXECUTE (due to FLUSH TABLES).
In such cases, the statement is re-prepared and re-executed from
Prepared_statement::execute_loop().

However, before the statement is reprepared, PXC's TOI_end() hook captures
this transient error (i.e, ER_NEED_REPREPARE) and sends it to galera thereby
triggering the inconsistenct voting and kicking this node out of the
cluster.

So, in such cases, we need to delay sending the final error to galera until
all reprepare attempts are over. This is done by attaching a temporary
Diagnostics_area to THD and resetting it later if all the required conditions
are met.