PXC-3113: Custom implementation of row hash calculation
https://perconadev.atlassian.net/browse/PXC-3113

When pxc_strict_mode is ENFORCING/MASTER, PXC prohibits writes to tables
without primary keys. This is the default setting.
When pxc_strict_mode is PERMISSIVE/DISABLED, writes to such tables
without PK is allowed. In such a case the row is identifed by MD5
checksum calculated from all columns values. OpenSSL library is used
for MD5 calculation.

On platform where FIPS mode is enabled, OpenSSL does not provide the
functionality of MD5 calculation. PXC could switch to use different
algorithm (like SHA), but that would mean that different version nodes
could not cooperate in the same cluster. MD5 is not used for
cryptography, only for hashing purpose (any algorithm would work as
long as it is the same on all nodes), so using it for this purpose is
FIPS compliant.

The solution is to fallback to custom implementation of MD5 hashing
for platforms where FIPS mode is enabled.

Unit tests introduced.
No new MTR tests introduced. We have several ones (galera_nopk_*.test)
that test usage of PXC without primary keys. When executed on FIPS
platform, they should be still OK.