PXC-4664: Executing transactions with the same GTID on different nodes causes the node to exit PXC-4688: Empty transaction stalls the replica thread
https://perconadev.atlassian.net/browse/PXC-4664
https://perconadev.atlassian.net/browse/PXC-4688

Both issues are the regression caused by fix for PXC-4664 (introduction
of wsrep_async_monitor).

Problem 1:
When node1 and node2 commit the transaction with the same gtid (after
setting 'gtid_next' to the same value), replicated node exits with
sigsegv.

Cause:
If the transaction with a given GTID was committed on local node, and
another, replicated transaction with the same GTID is being applied by
wsrep applier thread, we detect that this GTID was already committed and
attempt to skip the transaction
(gtid_pre_statement_checks() => skip_statement()).
In skip_statement() there is a code that marks a given transaction to be
skipped by wsrep_async_monitor. wsrep_async_monitor-related code is
valid only for async replica worker thread, not for wsrep applier.
We distinguish both by checking if thd is SYSTEM_THREAD_SLAVE_WORKER
(both galera and async worker) and 'wsrep_applier' flag. Check for
wsrep_applier was missing, so it was incorrectly detected, that we are
in the context of async worker thread. Then we did a dynamic_cast of
thd->rli_slave (of type Relay_log_info) to Slave_worker type.
Such a conversion gives a valid pointer only for async worker thread.
In case of wsrep applier, the object type is Relay_log_info, so
dynamic_cast results in nullptr. Then we dereference a pointer, and as
a consequence of being wsrep applier, we dereference nullptr.

Solution:
Added missing check for wsrep_applier, to access wsrep_async_monitor
path only from async worker thread.

For all similar places:
Added debug assertion.
Additionally, wrapped the block where we try to dereference a pointer
with a check against nullptr.

Note:
After fixing sigsegv, it is still possible to end up with a deadlock
(PXC-4665). It is an old issue.

Problem 2:
Committing an empty transaction on an async replica source causes the
PXC async replica node to stall.

Cause:
If an empty transaction is replicated from async source, PXC flow
commits empty transaction in Galera. This caused entering and releasing
wsrep_async_monitor. As this is an empty transaction, we need to commit
(and replicate) its GTID. It happens in
MYSQL_BIN_LOG::gtid_end_transaction(). Here we replicate and commit
an empty transaction with a given gtid. This path tried to acquire
wsrep_async_monitor again with the same sequence number, which caused
infinite wait.

Solution:
If we commit an empty transaction in Galera, but thd owns a GTID, it
means it will be commited later. In such a case postpone
entering/leaving releasing wsrep_async_monitor to the point when real
replication takes place.

Added MTR tests for both issues, however providing a reliable test for
PXC-4664 is not possible without fixing PXC-4665. PXC-4665 is an old
issue which causes a deadlock between wsrep applier and local session
(gtid lock vs galera commit_monitor).